[i3] cleanup, fix circle client

+++ i3/.config/i3/circle.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python3
-
+#!/usr/bin/env python3 
 """ i3 Named Scratchpads
 
 Usage:
@@ -17,14 +16,14 @@ Options:
 from docopt import docopt
 import os.path
 
-name_="circle"
-fifo_=os.path.realpath(os.path.expandvars('$HOME/tmp/'+name_+'.fifo'))
-
 if __name__ == '__main__':
+    name_="circle"
+    fifo_=os.path.realpath(os.path.expandvars('$HOME/tmp/'+name_+'.fifo'))
     argv = docopt(__doc__, version='i3 cycler over windows')
-    possible_commands=["next","info"]
+    if os.path.exists(fifo_):
+        possible_commands=["next","info"]
 
-    for i in argv:
-        if argv[i] and i in set(possible_commands):
-            with open(fifo_,"w") as fp:
-                fp.write(i+" "+argv["<name>"]+"\n")
+        for i in argv:
+            if argv[i] and i in set(possible_commands):
+                with open(fifo_,"w") as fp:
+                    fp.write(i+" "+argv["<name>"]+"\n")
